+++ ffi/src/scan.rs
@@ -13,7 +13,9 @@ use delta_kernel_ffi_macros::handle_descriptor;
 use tracing::debug;
 use url::Url;
 
-use crate::expressions::{unwrap_kernel_expression, EnginePredicate, KernelExpressionVisitorState};
+use crate::expressions::engine_visitor::{
+    unwrap_kernel_expression, EnginePredicate, KernelExpressionVisitorState,
+};
 use crate::{
     AllocateStringFn, ExclusiveEngineData, ExternEngine, ExternResult, IntoExternResult,
     KernelBoolSlice, KernelRowIndexArray, KernelStringSlice, NullableCvoid, SharedExternEngine,
