+++ ffi/src/test_ffi.rs
@@ -42,26 +42,26 @@ pub unsafe extern "C" fn get_testing_kernel_expression() -> Handle<SharedExpress
     .unwrap();
 
     let mut sub_exprs = vec![
-        Expr::literal(Scalar::Byte(i8::MAX)),
-        Expr::literal(Scalar::Byte(i8::MIN)),
-        Expr::literal(Scalar::Float(f32::MAX)),
-        Expr::literal(Scalar::Float(f32::MIN)),
-        Expr::literal(Scalar::Double(f64::MAX)),
-        Expr::literal(Scalar::Double(f64::MIN)),
-        Expr::literal(Scalar::Integer(i32::MAX)),
-        Expr::literal(Scalar::Integer(i32::MIN)),
-        Expr::literal(Scalar::Long(i64::MAX)),
-        Expr::literal(Scalar::Long(i64::MIN)),
+        Expr::literal(i8::MAX),
+        Expr::literal(i8::MIN),
+        Expr::literal(f32::MAX),
+        Expr::literal(f32::MIN),
+        Expr::literal(f64::MAX),
+        Expr::literal(f64::MIN),
+        Expr::literal(i32::MAX),
+        Expr::literal(i32::MIN),
+        Expr::literal(i64::MAX),
+        Expr::literal(i64::MIN),
         Expr::literal(Scalar::String("hello expressions".into())),
-        Expr::literal(Scalar::Boolean(true)),
-        Expr::literal(Scalar::Boolean(false)),
+        Expr::literal(true),
+        Expr::literal(false),
         Expr::literal(Scalar::Timestamp(50)),
         Expr::literal(Scalar::TimestampNtz(100)),
         Expr::literal(Scalar::Date(32)),
         Expr::literal(Scalar::Binary(0x0000deadbeefcafeu64.to_be_bytes().to_vec())),
         // Both the most and least significant u64 of the Decimal value will be 1
         Expr::literal(Scalar::Decimal((1 << 64) + 1, 2, 3)),
-        Expr::literal(Scalar::Null(DataType::Primitive(PrimitiveType::Short))),
+        Expr::literal(Scalar::Null(DataType::SHORT)),
         Expr::literal(Scalar::Struct(top_level_struct)),
         Expr::literal(Scalar::Array(array_data)),
         Expr::struct_expr(vec![Expr::or_from(vec![
